docs a bit
